network-script: add initial alive-check.sh script
This script helps to check if the DUT and LP are alive
and can have ssh session to check it has DUT|LP_REPORT.

If the alive-check fail, it automatically prints the
self-report logs under ~/test-logs.

If the target device does not self-report, then we need
to manually use the serial console to set GRUB to pick
GOLD-bzImage-kernel and once target machine is up,
at $HOME, run rescue_os.sh to revert the machine to use
GOLD-version of the kernel and modules

The GOLD-version of the kernel ingredients are setup
automatically if it does not exist during the step of
kernel upgrade steps.

Signed-off-by: Ong Boon Leong <elvinongbl@gmail.com>